my work dddd

+++ src/seller-page/SellerProductListing.jsx
@@ -10,10 +10,12 @@ const SellerProductListing = () => {
   }, [dispatch]);
 
   const handleDelete = (id) => {
-    dispatch(deleteProduct(id))
-      .unwrap()
-      .then(() => dispatch(getProduct()))
-      .catch((err) => console.log(err));
+    if (window.confirm('Do you really want to delete this item permanently?')) {
+      dispatch(deleteProduct(id))
+        .unwrap()
+        .then(() => dispatch(getProduct()))
+        .catch((err) => console.log(err));
+    }
   };
   return (
     <div>
@@ -21,20 +23,25 @@ const SellerProductListing = () => {
       <div>
         {sellerList?.map((item) => {
           return (
-            <div>
+            <div
+              className="flex justify-around items-center border hover:bg-slate-300"
+              key={item._id}
+            >
               <img
                 src={item.image}
                 alt=""
-                style={{ height: '200px', width: '200px' }}
+                style={{ height: '100px', width: '100px' }}
               />
+              <h2 className="w-48">{item.title}</h2>
+              <h2 className="w-48">Price:&#8377;{item.price}</h2>
               <button
-                className="btn bg-red-400"
+                className="btn bg-red-400 h-12"
                 onClick={() => handleDelete(item._id)}
               >
                 Delete
               </button>
               <Link to={`/seller/update-product/${item._id}`}>
-                <button className="btn bg-green-400">Update</button>
+                <button className="btn bg-green-400 p-2">Update</button>
               </Link>
             </div>
           );
